Modo vision (#70)
[OP#285](http://51.178.133.139:8080/projects/alliance-v0-dot-4/work_packages/285/activity)

Add modo vision mode inside Admin panel.
Modo vision allow to see any player names as if they were in the same
team.
This PR also include refactoring of "Open admin menu from selected user"
which now allow to enable the mouse and to choose the player we want.

This should ease the moderation.

# Toogle modo vision

Show all pseudo when pressing left ALT or not depending on modo vision
mode activated or not

# New "Open admin menu from selected user"

---------

Co-authored-by: Byako <byako.dev@outlook.fr>