Suggestion of refactoring on Messages declaration (#74)
I would suggest to move any msg into a separated msg static class in
order to centralize msg.
It would help to
- Add better structure to the app
- Easily know what messages propose an extension
- Avoid duplicate messages call definition
- Allow to easily see if message are used only inside an extensions or
used somewhere else (if this is the case, then some refactor may be
needed except if class is a core class for example)

This is currently only made on some of messages.
If approved, then we should refactor more messages in order to extract
them in any classes.