New menu for players spawn (#75)
### _**Still very much WIP**_

## Complete replacement of the native spawn menu for players

### Overall structure :
- a list of teams (with name, side - attacker or defender) => usually
not visible by clients
- for each team, a list of formations (with name, custom culture, and
various settings)
- for each formation, a list of available character (with customisable
slots and other various settings)

### Language :
Players can specify their preferred language for VOIP. This value is
stored in a local file and transferred to server when connecting or
changing the value.
Each formation displays the most common language among its members.

### Officer :
Each formation can have an officer. Formations members can see their
officer by pressing alt.
To become an officer, players need to select a compatible character in
the list (the ones with the red icon).
When selecting it, they will need to provide a pitch, and will become
candidates. Their candidacy will appear in the formation and the
formations members will be able to vote for them.
At the end of the election process (depends on game mode), the officer
with the most votes will be elected. The other candidates will be
dispatched to characters with available slots.

### Edition :
The menu can be edited by anyone when creating a new Scenario in the
Modding Kit
It can also be edited by admins : 
- at any point, during a game, by clicking on Edit mode. They can
load/save their local configurations and then click Sync to send their
configuration to the server, who will then share it with all clients.
- when launching a new game with the Game Mode menu, they can edit the
configuration that will be used on next game

### Currently implemented in :
- Lobby
- BattleX

### To do : 
- Compatibility with other game modes
- Handle officer disconnection / abandoning his post. Eventually provide
a way for players to demote their officer and start a new election ?
- Fix crappy perks UI, potentially adapt the native perk widget
- Limit characters animations, they can literally walk out of the screen
- Improve performance of the character preview, it's not good with a lot
of characters